Split handlers, readd report
