+++ .github/workflows/e2e.yml
@@ -36,10 +36,10 @@ jobs:
       - name: Run E2E tests
         run: npm run test:e2e
 
-      # - name: Upload Playwright report
-      #   if: always()
-      #   uses: actions/upload-artifact@v4
-      #   with:
-      #     name: playwright-report
-      #     path: playwright-report/
-      #     retention-days: 30
+      - name: Upload Playwright report
+        if: always()
+        uses: actions/upload-artifact@v4
+        with:
+          name: playwright-report
+          path: playwright-report/
+          retention-days: 30
